move package subspace to types
